+++ CHANGELOG.rst
@@ -2,6 +2,22 @@
 Change Log
 ==========
 
+
+-------------------
+v1.3.13 (2023-11-01)
+-------------------
+
+^^^^^^^^^
+Bug Fixes
+^^^^^^^^^
+* The default for the write_protected parameter when creating a master volume or master file system has changed from true to false. For snapshots, the default is true.
+
+^^^^^^^^^^^^^^^^^^^^
+Feature Enhancements
+^^^^^^^^^^^^^^^^^^^^
+* Add new infini_metadata module. This module will set, get and remove metadata (keys and values) to and from objects of these types: ["cluster", "fs", "fs-snap", "host", "pool", "system", "vol", "vol-snap"].
+* Add snapshot support to the infini_fs module. File system snapshot locks, regular and immutable are supported.
+
 -------------------
 v1.3.12 (2022-12-04)
 -------------------